# contratos/custodia.openapi.yaml
openapi: 3.1.0
info:
  title: Custodia documental (MS-04)
  version: 0.9.0
  description: >
    Documentos del ciudadano afiliado. Realiza HU-02 (RF-09.1, RNF-10, RNF-11) y HU-03
    (RF-02.2, RF-02.3, RNF-24, RI-06) y HU-04 (RF-06.6, RF-02.5, RI-01, RNF-21) y HU-05 (RF-03.2, RF-03.3) y HU-06 (RF-02.7, RNF-01) y HU-07 (RI-08) y HU-08 (RF-04.1) y HU-09 (RF-03.3, RF-03.5, RNF-22): todas las rutas
    salvo /salud exigen un Bearer JWT del emisor OIDC con audiencia `custodia` y `azp: portal`;
    sin firma válida, de otro emisor, de otra audiencia o vencido responde 401. Cada ciudadano
    solo ve su Carpeta. El binario nunca pasa por este servicio: sube directo al almacén por una
    URL prefirmada de 5 minutos (RI-06) y la custodia solo recibe metadatos (cuerpos de 2 KB).
    Las rutas `/interno/*` son de la interoperabilidad (MS-07): exigen un token de servicio
    (client credentials, `azp: interoperabilidad`); un token de ciudadano recibe 403.
servers:
  - url: http://localhost:8083
security:
  - bearer: []
paths:
  /documentos:
    get:
      operationId: listarDocumentos
      summary: Lista los documentos Cargados de la Carpeta del titular
      description: >
        Lectura directa de la custodia. La SPA consulta y busca en el índice de carpeta (MS-05,
        `indice.openapi.yaml`); esta ruta queda para la conciliación.
      responses:
        '200':
          description: Documentos del titular, del más reciente al más antiguo
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Documento' }
        '401': { $ref: '#/components/responses/Problema' }
    post:
      operationId: registrarDocumento
      summary: Registra un Temporal pendiente y entrega la URL PUT prefirmada de 5 minutos
      description: >
        Valida título, formato (PDF, JPG o PNG), tamaño (10 MB) y cuota (20 Temporales y 200 MB)
        antes de emitir la URL. No hay clave por documento: el objeto se llama `{cédula}/{id}`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [titulo, tipo, tamano]
              properties:
                titulo: { type: string, minLength: 1, maxLength: 120 }
                tipo: { type: string, enum: [application/pdf, image/jpeg, image/png] }
                tamano: { type: integer, minimum: 1, maximum: 10485760, description: Bytes }
      responses:
        '201':
          description: Pendiente de carga
          content:
            application/json:
              schema:
                type: object
                required: [id, urlCarga]
                properties:
                  id: { type: string, format: uuid }
                  urlCarga: { type: string, format: uri }
        '401': { $ref: '#/components/responses/Problema' }
        '413': { $ref: '#/components/responses/Problema' }
        '415': { $ref: '#/components/responses/Problema' }
        '422': { $ref: '#/components/responses/Problema' }
  /documentos/{id}/confirmacion:
    post:
      operationId: confirmarCarga
      summary: Verifica el objeto subido y deja el documento Cargado
      description: >
        Compara tamaño y tipo con lo declarado y calcula el SHA-256 del objeto. Si no coinciden
        borra el objeto y responde 422. Es idempotente: confirmar un Cargado lo devuelve igual.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Documento Cargado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Documento' }
        '401': { $ref: '#/components/responses/Problema' }
        '403': { $ref: '#/components/responses/Problema' }
        '422': { $ref: '#/components/responses/Problema' }
...
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer, bearerFormat: JWT }
    servicio: { type: http, scheme: bearer, bearerFormat: JWT, description: Token client credentials de la interoperabilidad }
  schemas:
    Documento:
      type: object
      required: [id, titulo, clase, estado, tipo, tamano, creado]
      properties:
        id: { type: string, format: uuid }
        titulo: { type: string }
        clase: { type: string, enum: [temporal, certificado] }
        estado:
          type: string
          description: >
            Temporal: cargado, sustituido, eliminado. Certificado: recibido, verificado, vigente,
            rechazado, retirado. La Carpeta del titular solo muestra Temporales cargado o sustituido
            y Certificados vigente.
          enum: [cargado, sustituido, eliminado, recibido, verificado, vigente, rechazado, retirado]
        emisor: { type: string, description: Entidad que firmó el Certificado }
        sustituidoPor: { type: string, format: uuid, description: Certificado que reemplazó a este Temporal }
        motivo: { type: string, description: Por qué quedó Rechazado el Certificado; solo en las rutas internas }
        titular: { type: string, description: Cédula del titular; solo en las rutas internas }
        tipo: { type: string, enum: [application/pdf, image/jpeg, image/png] }
        tamano: { type: integer }
        sha256: { type: string, description: Huella del archivo calculada al confirmar la carga }
        autenticacion:
          type: object
          description: Constancia de GovCarpeta; ausente mientras el Temporal no esté Autenticado
          required: [fecha, respuesta]
          properties:
            fecha: { type: string, format: date-time }
            respuesta: { type: string, description: 'Texto del centralizador: un sello suyo, no un certificado (B-03)' }
        creado: { type: string, format: date-time }
...
  responses:
    Problema:
      description: Error RFC 9457
      content:
        application/problem+json:
          schema:
            type: object
            required: [type, title, status]
            properties:
              type: { type: string }
              title: { type: string }
              status: { type: integer }
              detail: { type: string }